set layer dashboards
