+++ src/views/dashboard/graphs/ServiceList.vue
@@ -20,7 +20,7 @@ limitations under the License. -->
         placeholder="Please input service name"
         size="small"
         @change="searchList"
-        class="inputs"
+        class="inputs mt-5"
       >
         <template #append>
           <el-button size="small" @click="searchList">
